pyproject.toml
[tool.poetry]
name = "mipi_env_manager"
version = "0.3.0"
description = ""
authors = ["Sean Murphy <murphy-sean1@cooperhealth.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
packaging = "^24.2"
pyyaml = "^6.0.2"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
jinja2 = "^3.1.5"
click = "^8.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mipi-build-envs = "mipi_env_manager.main:mipi-publish-envs"
